Ignore reflection calls inside reflection implementation
Some relfection methods are implemented by calling different overloads of the same method. In these cases linker should not try to analyze the internal call.

For example:
```C#
MethodInfo GetMethod(string name) => GetMethod(name, null);
```
In this case the call to `GetMethod(name, null)` should not be analyzed and should not cause "warnings".

Added new test infra to be able to capture and validate logged messages.